+++ CHANGELOG.md
@@ -53,6 +53,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## Log
 
+## [version] - @todo
+
+### Added
+
+* Convert path string to posix format
+* Convert path string to win32 format
+
 ## [1.2.0] - 2021-05-28
 
 ### Added